Update Readme:  remove "pure" from pure function
https://www.reddit.com/r/reactjs/comments/abj7hn/uis_are_not_pure_functions_of_the_model_reactjs/ed1h20x/

"To elaborate a bit, React components aren’t always “pure” in strict FP sense of the word. They’re also not always functions (although we’re adding a stateful function API as a preferred alternative to classes soon). Support for local state and side effects is absolutely a core feature of React components and not something we avoid for “purity"